+++ tools/njsq/wrapper.cpp
@@ -19,49 +19,11 @@
 
 #include "wrapper.h"
 
+#include "n4aups.h"
+#include "sclerror.h"
 #include "v8q.h"
 
-#include "sclerror.h"
-
 using namespace wrapper;
-
-namespace {
-	using n4a::cLauncher;
-
-	cLauncher *Launcher_ = NULL;
-
-	cLauncher &GetLauncher_( void )
-	{
-		if ( Launcher_ == NULL )
-			qRGnr();
-
-		return *Launcher_;
-	}
-}
-
-void wrapper::SetLauncher( cLauncher *Launcher )
-{
-	if ( Launcher_ != NULL )
-		qRGnr();
-
-	if ( Launcher == NULL )
-		qRGnr();
-
-	Launcher_ = Launcher;
-}
-
-bso::sBool wrapper::GetLauncherInfo( str::dString &Info )
-{
-	if ( Launcher_ == NULL ) {
-		sclmisc::GetBaseTranslation( common::message::NoRegisteredComponent, Info );
-		return false;
-	}
-
-	Launcher_->Info( Info );
-
-	return true;
-}
-
 
 typedef n4a::cCaller cCaller_;
 
@@ -139,9 +101,9 @@ namespace {
 
 void wrapper::Launch( const v8::FunctionCallbackInfo<v8::Value>& Info )
 {
-qRFH
+qRH
 	sCaller_ Caller;
-qRFB
+qRB
 	if ( Info.Length() < 1 )
 		qRGnr();
 
@@ -150,22 +112,14 @@ qRFB
 
 	v8::Local<v8::Uint32> Index = v8::Local<v8::Uint32>::Cast(Info[0] );
 
-	if ( !common::Functions.Exists( Index->Uint32Value() ) )
-		qRGnr();
-
 	Caller.Init( Info );
 
-	GetLauncher_().Launch( common::Functions( Index->Uint32Value()), Caller );
+	n4aups::GetLauncher().Launch( n4aups::GetFunction( Index->Uint32Value()), Caller );
 
 	if ( sclerror::IsErrorPending() )
 		qRAbort();	// To force the handling of a pending error.
-qRFR
-qRFT
-qRFE( common::ErrFinal( Info.GetIsolate() ) )
+qRR
+qRT
+qRE
 }
 
-void wrapper::DeleteLauncher( void )
-{
-	if ( Launcher_ != NULL )
-		delete Launcher_;
-}