+++ tools/njsq/CHANGELOG.md
@@ -1,4 +1,7 @@
-# NJSq changelog
+# *NJSq* changelog
+
+2017-08-10 :
+- Some code is moved to *N4AUPS* library.
 
 2017-08-09 :
 - Using *N4A* library instead of *NJS* one.